+++ packages/eve/src/execution/dispatch-runtime-actions-step.ts
@@ -31,14 +31,8 @@ import {
   type RuntimeAgentHandleAction,
   type RuntimeSession,
 } from "#execution/agent-handle-dispatch.js";
-import { REMOTE_AGENT_START_FAILED, SUBAGENT_START_FAILED } from "#harness/agent-handle-errors.js";
-import { deriveAgentOperationId } from "#harness/handles/operation-id.js";
-import {
-  deriveAgentId,
-  getAgentHandleStore,
-  type AgentIdentity,
-  type StartOperation,
-} from "#harness/handles/store.js";
+import { SUBAGENT_START_FAILED } from "#harness/agent-handle-errors.js";
+import { getAgentHandleStore } from "#harness/handles/store.js";
 import {
   confirmAgentStarted,
   prepareAgentStart,
@@ -66,6 +60,13 @@ import {
   resolveRemoteAgentForAction,
   startRemoteAgentSession,
 } from "#execution/remote-agent-dispatch.js";
+import {
+  createRecursiveAgentRootOnlyResult,
+  createRemoteAgentStartFailureResult,
+  createUnavailableDynamicSubagentResult,
+  getSubagentName,
+} from "#execution/dispatch-action-failures.js";
+import { mintStartOperation } from "#execution/dispatch-start-operation.js";
 import { hydrateDurableSession } from "#execution/session.js";
 import { buildSubagentRunInput, type SubagentInputSource } from "#execution/subagent-tool.js";
 import { createWorkflowRuntime, workflowEntryReference } from "#execution/workflow-runtime.js";
@@ -464,38 +465,6 @@ async function startSubagent(input: {
   }
 }
 
-/**
- * Mints the deterministic start operation and identity for one dispatch.
- * All inputs are parent-controlled, so both exist before the child does
- * and a durable replay of the step re-derives the same ownership record.
- */
-function mintStartOperation(input: {
-  readonly callId: string;
-  readonly name: string;
-  readonly nodeId: string;
-  readonly parentSessionId: string;
-  readonly parentTurnId: string;
-}): { readonly identity: AgentIdentity; readonly operation: StartOperation } {
-  const operationId = deriveAgentOperationId({
-    callId: input.callId,
-    parentSessionId: input.parentSessionId,
-    parentTurnId: input.parentTurnId,
-  });
-  return {
-    identity: {
-      id: deriveAgentId(input.name, operationId),
-      name: input.name,
-      nodeId: input.nodeId,
-    },
-    operation: {
-      callId: input.callId,
-      id: operationId,
-      kind: "start",
-      parentTurnId: input.parentTurnId,
-    },
-  };
-}
-
 async function startLocalSubagent(input: {
   readonly action: RuntimeSubagentCallActionRequest;
   readonly auth: Parameters<typeof buildSubagentRunInput>[0]["auth"];
@@ -703,62 +672,6 @@ async function startRemoteSubagent(input: {
       }),
     };
   }
-}
-
-function createUnavailableDynamicSubagentResult(
-  action: RuntimeSubagentCallActionRequest | RuntimeRemoteAgentCallActionRequest,
-): RuntimeSubagentDispatchFailure {
-  const subagentName = getSubagentName(action);
-  return {
-    callId: action.callId,
-    isError: true,
-    kind: "subagent-result",
-    origin: "dispatch",
-    output: {
-      code: "SUBAGENT_UNAVAILABLE",
-      message: `Subagent "${subagentName}" is not available in the current session context.`,
-    },
-    subagentName,
-  };
-}
-
-function getSubagentName(
-  action: RuntimeSubagentCallActionRequest | RuntimeRemoteAgentCallActionRequest,
-): string {
-  return action.kind === "remote-agent-call" ? action.remoteAgentName : action.subagentName;
-}
-
-function createRemoteAgentStartFailureResult(input: {
-  readonly action: RuntimeRemoteAgentCallActionRequest;
-  readonly error: unknown;
-}): RuntimeSubagentDispatchFailure {
-  return {
-    callId: input.action.callId,
-    isError: true,
-    kind: "subagent-result",
-    origin: "dispatch",
-    output: {
-      code: REMOTE_AGENT_START_FAILED,
-      message: toErrorMessage(input.error),
-    },
-    subagentName: input.action.remoteAgentName,
-  };
-}
-
-function createRecursiveAgentRootOnlyResult(
-  action: RuntimeSubagentCallActionRequest,
-): RuntimeSubagentDispatchFailure {
-  return {
-    callId: action.callId,
-    isError: true,
-    kind: "subagent-result",
-    origin: "dispatch",
-    output: {
-      code: "RECURSIVE_AGENT_ROOT_ONLY",
-      message: 'The built-in "agent" tool is only available to the root session.',
-    },
-    subagentName: action.subagentName,
-  };
 }
 
 function isRecursiveAgentAction(
